+++ portal/apps/homev3/static/js/homev2/dynamic_edition_admin.js
@@ -5,30 +5,31 @@ var InlineOrdering = {
         $('td[class$=-top_position]').hide();
         $('td[class$=-position]').hide();
 
+        var inline_group_containers = $('.inline-group').has('.dynamic-order');
         // Rompemos el tbody para hacer multiples tbody.
-        $('.inline-group table > tbody').contents().unwrap();
+        $('.inline-group > .dynamic-order > table > tbody').contents().unwrap();
 
         var MakePages = function(section){
             var section = $(section);
             var section_id = section.data('section_id');
-            var articles = section.find('table tr[id^=articlerel_set]').not('[id$=-empty]');
+            var articles = section.find('.dynamic-order > table tr').not('[id$=-empty]');
             $(articles).each( function(i, art){
                 var art = $(art);
                 var new_page = art.find("input[id$=-page]").val();
-                var page = art.closest('table').find('tbody[data-page="' + new_page + '"]');
+                var page = art.closest('.dynamic-order > table').find('tbody[data-page="' + new_page + '"]');
                 var exists = page.length != 0;
                 if (page != new_page || page == 'undefined'){
-                    var tbody = art.closest('table').find('tbody[data-page="' + new_page + '"]');
+                    var tbody = art.closest('.dynamic-order > table').find('tbody[data-page="' + new_page + '"]');
                     var exists = tbody.length != 0;
-                    var articles_count = tbody.find('table tr[id^=articlerel_set]').not('[id$=-empty]');
+                    var articles_count = tbody.find('.dynamic-order > table tr').not('[id$=-empty]');
                     if (exists && articles_count.length < 6){
                         tbody.append(art);
                     }else{
                         if (articles_count.length > 6){
                             new_page = new_page + "1";
                         }
-                        art.closest('table').append('<tbody class="page page_' + new_page + '" data-page="' + new_page + '"></tbody>');
-                        parent_tbody = art.closest('table').find('tbody[data-page="' + new_page + '"]');
+                        art.closest('.dynamic-order > table').append('<tbody class="page page_' + new_page + '" data-page="' + new_page + '"></tbody>');
+                        parent_tbody = art.closest('.dynamic-order > table').find('tbody[data-page="' + new_page + '"]');
                         parent_tbody.append(art);
                         art.attr('data-page', new_page);
                     }
@@ -37,7 +38,7 @@ var InlineOrdering = {
         };
 
         function UpdateArticlesPage(section){
-            var articles = $(section).find('tr[id^=articlerel_set]').not('[id$=-empty]');
+            var articles = $(section).find('.dynamic-order > table > tr').not('[id$=-empty]');
             articles.each(function(i,art){
                 var art = $(art);
                 var input_page = art.find('input[id$=-page]');
@@ -47,7 +48,7 @@ var InlineOrdering = {
         };
 
         function UpdatePages(section){
-            var table = $(section).find('table');
+            var table = $(section).find('.dynamic-order > table');
             var pages = table.find('> tbody');
             pages.each(function(i,page){
                 var page = $(page);
@@ -75,7 +76,7 @@ var InlineOrdering = {
             var section = $(this).closest('.inline-group');
             var new_page = $(this).val();
             var current_page = $(this).closest('tbody').data('page');
-            var arts_count = section.find('tbody[data-page="' + new_page + '"]').find('tr[id^=articlerel_set]').not('[id$=-empty]').length;
+            var arts_count = section.find('.dynamic-order > table > tbody[data-page="' + new_page + '"]').find('tr').not('[id$=-empty]').length;
 
             if (arts_count > 5){
                 $(this).val(current_page);
@@ -84,9 +85,9 @@ var InlineOrdering = {
             }
         });
 
-        $('div[id^=articlerel_set-].inline-group').not('[id=articlerel_set-group]').each( function(i, section){
+        inline_group_containers.not('[id$=-group]').each( function(i, section){
             var section = $(section);
-            var orderables = section.find('[id^=articlerel_set]');
+            var orderables = section.find('.dynamic-order > table > tbody > tr').has('td');
             var connector = '.page';
 
             $('.placeholder').hide();
@@ -104,8 +105,8 @@ var InlineOrdering = {
                     UpdateSection(section);
                 },
                 update: function(event, ui) {
-                    page = ui.item.closest('tbody');
-                    if (page.find('tr[id^=articlerel_set]').not('[id$=-empty]').length > 6 && connector != ''){
+                    page = ui.item.closest('.dynamic-order > table > tbody');
+                    if (page.find('tr').not('[id$=-empty]').length > 6 && connector != ''){
                         section.sortable('cancel');
                     }
                 },
@@ -114,9 +115,9 @@ var InlineOrdering = {
             $('td.field-order input').hide();
         });
 
-        $('#articlerel_set-group').each( function(){
+        inline_group_containers.each( function(){
             var section = $(this);
-            var orderables = section.find('[id^=articlerel_set]');
+            var orderables = section.find('.dynamic-order > table > tbody > tr').has('td');
             section.sortable({
                 axis: 'y',
                 placeholder: 'ui-state-highlight',
@@ -138,10 +139,31 @@ var InlineOrdering = {
         });
 
         $('form').submit(function(){
-            $('.inline-group').each( function(){
+            inline_group_containers.each( function(){
+                var container = $(this);
                 var i = 0;
-                $(this).find('tr[id^=articlerel_set]').not('tr[id$=-empty]').each( function(){
-                    var order_input = $(this).find('input[id$=position]');
+                if(container.find('tr[id^=articlerel_set]').length > 0){
+                    var order_field = 'top_position';
+                    // For fix the entire page uncomment these lines and build the section for store
+                    /*if(container.find('h2:contains("Nota de tapa y titulines")').length > 0){
+                        var order_field = 'top_position';
+                    } else {
+                        var order_field = 'position';
+                    }*/
+                    var tr_elem = container.find('tr[id^=articlerel_set]').filter('.has_original');
+                }
+                else if(container.find('tr[id^=categorynewsletterarticle_set]').length > 0){
+                    var order_field = 'order';
+                    var tr_elem = container.find('tr[id^=categorynewsletterarticle_set]').filter('.has_original');
+                }
+                else{
+                    var order_field = '#';  // not matched character
+                    var tr_elem = $(this);
+                }
+
+                tr_elem.not('tr[id$=-empty]').each( function(){
+                    var input_find = ['input[id$=', order_field, ']'].join('');
+                    var order_input = $(this).find(input_find);
                     // Pone en 0 antes del submit si no tiene article
                     if ($(this).find('input[id$=headline]').val() == ''){
                         order_input.val(0);
@@ -157,7 +179,8 @@ var InlineOrdering = {
 
 var SectionIds = {
     init: function () {
-        $('div[id^=articlerel_set-].inline-group').not('[id=articlerel_set-group]').each(function(){
+        var inline_group_containers = $('.inline-group').has('.dynamic-order');
+        inline_group_containers.not('[id$=-group]').each(function(){
             var div = $(this);
             var title = $(div.find('h2'));
             var text = title.html();
@@ -176,7 +199,16 @@ var PopUpAdd = {
     init: function () {
         $('div.add-row a').each(function(){
             var add_link = $(this);
-            var section_id = add_link.closest('div[id^=articlerel_set-].inline-group').attr('data-section_id');
+            if(add_link.closest('div[id^=articlerel_set-].inline-group').length > 0){
+                var cover_group = add_link.closest('div[id^=articlerel_set-].inline-group');
+            }
+            else if (add_link.closest('div[id^=categorynewsletterarticle_set-].inline-group').length > 0){
+                var cover_group = add_link.closest('div[id^=categorynewsletterarticle_set-].inline-group');
+            }
+            else {
+                var cover_group = undefined;
+            }
+            var section_id = cover_group.attr('data-section_id');
             var pub_id = $('#id_publication').val();
             add_link.off("click");
             add_link.on("click", function(){
@@ -204,16 +236,31 @@ function dismissAddAnotherPopup(win, newId, newRepr) {
     location.reload();
 };
 
+function dismissRelatedLookupPopupTarget(win, chosenId) {
+        /*
+        Convert open window name into an django html element ID format.
+        Find element by ID if it's vManyToManyRawIdAdminField attach given ID to the value.
+        Otherwise, set the given ID to the value of the target element
+        */
+        var name = windowname_to_id(win.name);
+        var elem = window.document.getElementById(name);
+        if (elem){
+            if (elem.className.indexOf('vManyToManyRawIdAdminField') !== -1 && elem.value) {
+                elem.value += ',' + chosenId;
+            } else {
+                var is_empty = (elem.value) ? false: true;
+                elem.value = chosenId;
+                // add class to the parent tr for prevent default order for the new row added, and keep the selected order
+                if(is_empty){
+                    $('#' + name).closest('tr').addClass('has_original');
+                }
+            }
+        }
+        win.close();
+    }
+
 function dismissRelatedLookupPopup(win, article_id){
-    var last_inline = last_clicked_inline.closest('tbody').prev().prev();
-    var new_inline = last_inline.clone();
-    last_inline.after(new_inline);
-    var a = $('a[href="' + article_id + '/"]', win.document);
-    var headline = a.html();
-    new_inline.find('input[id$=headline]').val(headline);
-    var edition_id = $('#id_articlerel_set-0-edition').val();
-    win.close();
-    window.location.href = '/admin/core/edition/add_article/' + edition_id + '/' + last_section_id + '/' + article_id + '/';
+    dismissRelatedLookupPopupTarget(win, article_id);
 }
 
 $(function () {