+++ portal/migration_settings.py
@@ -116,6 +116,7 @@ INSTALLED_APPS = (
     'dashboard',
     'django_filters',
     'rest_framework',
+    'rest_framework.authtoken',
     'compressor',
     'favit',
     'social_django',
@@ -156,9 +157,15 @@ ADMIN_SHORTCUTS = [
 ]
 
 REST_FRAMEWORK = {
-    'DEFAULT_PERMISSION_CLASSES': ('rest_framework.permissions.IsAdminUser',),
-    'PAGINATE_BY': 20,
-    'DEFAULT_FILTER_BACKENDS': ('django_filters.rest_framework.DjangoFilterBackend',),
+    'DEFAULT_PERMISSION_CLASSES': ('rest_framework.permissions.IsAdminUser', ),
+    'DEFAULT_AUTHENTICATION_CLASSES': (
+        'rest_framework.authentication.BasicAuthentication',
+        'rest_framework.authentication.SessionAuthentication',
+        'rest_framework.authentication.TokenAuthentication',
+    ),
+    'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.LimitOffsetPagination',
+    'PAGE_SIZE': 20,
+    'DEFAULT_FILTER_BACKENDS': ('django_filters.rest_framework.DjangoFilterBackend', ),
 }
 
 ACTSTREAM_SETTINGS = {'FETCH_RELATIONS': False, 'USE_PREFETCH': True}
@@ -377,6 +384,7 @@ CORE_ENABLE_RELATED_ARTICLES = True
 
 # mongodb database
 MONGODB_DATABASE = 'utopia_cms'
+MONGODB_NOTIMEOUT_CURSORS_ALLOWED = True
 
 # Change to false if the signupwall middleware is removed
 SIGNUPWALL_ENABLED = True